Refactor query results processing
